Fix export config for tools - Treat 'vue' dependency as a webpack external for tools builds - Split main webpack config into build and dev(serve) configs
