--- tools/api/fastApi/index.js
@@ -1,2 +1,0 @@
-export { fastApiTableAdapter } from './adapter';
-export { fastApiTableAdapterComposable } from './vue-composable';
